cpufreq: remove -Wabsolute-value warning cpufreq-selector-sysfs.c:244:18: warning: taking the absolute value of unsigned type ‘guint’ {aka ‘unsigned int’} has no effect [-Wabsolute-value]   244 |   current_dist = abs (frequency - f);       |                  ^~~ -- cpufreq-selector-procfs.c:187:7: warning: taking the absolute value of unsigned type ‘guint’ {aka ‘unsigned int’} has no effect [-Wabsolute-value]   187 |   if (abs (sc_max - frequency) < abs (frequency - sc_min))       |       ^~~ cpufreq-selector-procfs.c:187:34: warning: taking the absolute value of unsigned type ‘guint’ {aka ‘unsigned int’} has no effect [-Wabsolute-value]   187 |   if (abs (sc_max - frequency) < abs (frequency - sc_min))       |                                  ^~~ -- cpufreq-selector-libcpufreq.c:109:18: warning: taking the absolute value of unsigned type ‘long unsigned int’ has no effect [-Wabsolute-value]   109 |   current_dist = abs (freq->frequency - frequency);       |                  ^~~

+++ cpufreq/src/cpufreq-selector/cpufreq-selector-procfs.c
@@ -184,7 +184,7 @@ cpufreq_selector_procfs_set_frequency (CPUFreqSelector *selector,
 	}
 
 	if (frequency != sc_max && frequency != sc_min) {
-		if (abs (sc_max - frequency) < abs (frequency - sc_min))
+		if (abs ((int)sc_max - (int)frequency) < abs ((int)frequency - (int)sc_min))
 			freq = sc_max;
 		else
 			freq = sc_min;
